Attempt to refactor code in multiple files
