+++ utils.go
@@ -1,0 +1,36 @@
+package main
+
+import (
+	"flag"
+	"fmt"
+	"log"
+
+	"github.com/fatih/color"
+	"github.com/pkg/errors"
+)
+
+func checkf(err error, format string, args ...interface{}) {
+	if err != nil {
+		log.Printf(format, args...)
+		log.Println()
+		log.Fatalf("%+v", errors.WithStack(err))
+	}
+}
+
+func assertf(ok bool, format string, args ...interface{}) {
+	if !ok {
+		log.Printf(format, args...)
+		log.Println()
+		log.Fatalf("%+v", errors.Errorf("Should be true, but is false"))
+	}
+}
+
+var errc = color.New(color.BgRed, color.FgWhite).PrintfFunc()
+
+func oerr(msg string) {
+	errc("\tERROR: " + msg + " ")
+	fmt.Println()
+	fmt.Println("Flags available:")
+	flag.PrintDefaults()
+	fmt.Println()
+}
